Yikes. Mismatched date indexes. Label colors. IsOpenNow.
Fixed the date indexes so they match. Swift calendar week is 1-7. Sunday-Saturday. Yelp calendar week is 0-6 Monday-Sunday.
Labels were white so it looked like it was broken... That took longer than I care to admit to realize.
is_closed now pointing to is_open_now in json data. is_closed is if the business is PERMANENTLY closed or not.... going to use that later to scrub the initial list since  I don't even want to list restaraunts that are PERMANENTLY CLOSED....